rename component as "ButtonGroup" since it doesn't handle its own toggle state
