+++ web-common/src/components/button-group/ButtonGroup.svelte
