+++ web-common/src/components/button-group/index.ts
@@ -1,4 +1,4 @@
 /** See README.md for more information. */
 
-export { default as ButtonToggleGroup } from "./ButtonToggleGroup.svelte";
+export { default as ButtonGroup } from "./ButtonGroup.svelte";
 export { default as SubButton } from "./SubButton.svelte";
